+++ lib/eve_online.rb
@@ -48,6 +48,7 @@ require 'eve_online/esi/market_history'
 
 require 'eve_online/esi/corporation'
 require 'eve_online/esi/corporation_industry_jobs'
+require 'eve_online/esi/corporation_blueprints'
 
 require 'eve_online/esi/alliance'
 require 'eve_online/esi/alliance_icon'